Fix warning
[SVN r2975]

+++ include/boost/math/concepts/real_concept.hpp
@@ -240,7 +240,7 @@ inline float real_cast<float, concepts::real_concept>(concepts::real_concept r)
 template <>
 inline double real_cast<double, concepts::real_concept>(concepts::real_concept r)
 {
-   return r.value();
+   static_cast<double>(return r.value());
 }
 template <>
 inline long double real_cast<long double, concepts::real_concept>(concepts::real_concept r)
